+++ src/conf/serverconf.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   serverconf.cpp                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: emadriga <emadriga@student.42madrid.com>   +#+  +:+       +#+        */
+/*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/20 20:28:53 by jvacaris          #+#    #+#             */
-/*   Updated: 2023/05/23 15:43:25 by emadriga         ###   ########.fr       */
+/*   Updated: 2023/05/23 16:38:09 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,3 +22,9 @@ ft::serverconf::~serverconf()
 	if(locations.size())
 		locations.clear();
 }
+
+ft::serverconf::serverconf(const ft::serverconf & tocopy): address(tocopy.address), port(tocopy.port), default_root(tocopy.default_root), locations(tocopy.locations), filetypes(tocopy.filetypes)
+{
+
+}
+