+++ src/conf/serverconf.hpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   serverconf.hpp                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: emadriga <emadriga@student.42madrid.com>   +#+  +:+       +#+        */
+/*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/01 17:15:18 by emadriga          #+#    #+#             */
-/*   Updated: 2023/05/23 15:46:46 by emadriga         ###   ########.fr       */
+/*   Updated: 2023/05/23 16:37:42 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,15 +46,11 @@ class serverconf{
 		~serverconf();
 
 		//Copy constructor
-	private:
-		//serverconf musn't be updated after creation readonly copy or assign would worse performance
-		serverconf();
-		serverconf( const serverconf& copy );
-		serverconf & operator=( const serverconf& assign );
-
+		serverconf(const ft::serverconf & tocopy);
 
 };
 
 }//Namespace ft
 
 #endif
+